Refactor: Create a deduplication file
