+++ specifyweb/specify/migration_utils/SchemaReader.py
@@ -15,6 +15,7 @@ from django.core.exceptions import MultipleObjectsReturned
 from django.db import connection, transaction
 from django.db.models.functions import RowNumber
 
+from specifyweb.specify.migration_utils.SchemaWriter import FieldDefaults
 from specifyweb.specify.models_utils.load_datamodel import FieldDoesNotExistError, TableDoesNotExistError
 
 from specifyweb.specify.models_utils.load_datamodel import Table, FieldDoesNotExistError, TableDoesNotExistError
